models/convert_int8/requirements.txt
torchvision
tqdm
scikit-learn
fastdownload
openvino-dev==2023.2
